Move exception types to flux-common, and add a common base class for them.
https://github.com/danielgmyers/flux-swf-client/issues/129
