+++ flux-common/src/main/java/com/danielgmyers/flux/ex/BadWorkflowStateException.java
@@ -1,0 +1,29 @@
+/*
+ *   Copyright Flux Contributors
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License").
+ *   You may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+package com.danielgmyers.flux.ex;
+
+/**
+ * Indicates that the current state of the workflow has caused Flux to be unable to determine what it should do next.
+ * For example, if a step returns an unknown result code, Flux may be unable to determine what step should run next.
+ */
+public class BadWorkflowStateException extends FluxException {
+
+    public BadWorkflowStateException(String message) {
+        super(message);
+    }
+
+}
