+++ flux-common/src/main/java/com/danielgmyers/flux/ex/FluxException.java
@@ -14,19 +14,18 @@
  *   limitations under the License.
  */
 
-package com.danielgmyers.flux.clients.swf.poller;
+package com.danielgmyers.flux.ex;
 
 /**
- * Indicates that a task was scheduled that the activity worker does not know about.
- * Typically this will happen during deployment after adding new steps;
- * if it happens after removing a step (or rolling back a deployment)
- * the affected workflow will be stuck until terminated or the activity
- * is made available again.
+ * A common base class for exceptions generated by Flux code.
  */
-public class UnrecognizedTaskException extends RuntimeException {
+public class FluxException extends RuntimeException {
 
-    public UnrecognizedTaskException(String message) {
+    public FluxException(String message) {
         super(message);
     }
 
+    public FluxException(String message, Throwable cause) {
+        super(message, cause);
+    }
 }
