+++ flux-swf/src/main/java/com/danielgmyers/flux/clients/swf/poller/WorkflowState.java
@@ -26,6 +26,7 @@ import java.util.Map;
 import java.util.Set;
 
 import com.danielgmyers.flux.clients.swf.poller.timers.TimerData;
+import com.danielgmyers.flux.ex.BadWorkflowStateException;
 import com.danielgmyers.flux.poller.TaskNaming;
 import com.danielgmyers.flux.signals.BaseSignalData;
 import com.danielgmyers.flux.signals.ForceResultSignalData;
