+++ flux-swf/src/test/java/com/danielgmyers/flux/clients/swf/RemoteWorkflowExecutorTest.java
@@ -24,6 +24,7 @@ import java.util.Set;
 
 import com.danielgmyers.flux.clients.swf.poller.testwf.TestWorkflow;
 import com.danielgmyers.flux.clients.swf.poller.testwf.TestWorkflowWithPartitionedStep;
+import com.danielgmyers.flux.ex.WorkflowExecutionException;
 import com.danielgmyers.flux.poller.TaskNaming;
 import com.danielgmyers.flux.testutil.ManualClock;
 import com.danielgmyers.flux.wf.Workflow;
